# swagger-doc.yaml
openapi: 3.0.0
info:
  title: FitAura Full API Collection
  version: 1.0.0
paths:
...
  /nutrition-logs:
    post:
      summary: Add meal log
      responses:
        '200':
          description: Successful response
      requestBody:
        required: true
        content:
          application/json:
            example:
              user_id: 1
              date: '2025-04-15'
              meal_type: lunch
              food_item: Grilled Chicken
              calories: 450
              protein: 35
              carbs: 10
              fat: 20
...
  /progress-logs:
    post:
      summary: Log new entry
      responses:
        '200':
          description: Successful response
      requestBody:
        required: true
        content:
          application/json:
            example:
              user_id: 2
              date: '2025-04-15'
              weight: 95.6
              body_fat: 20.5
              notes: Lost 1kg this week
...